+++ lib/ts/recipe/thirdparty/components/themes/signInAndUp/index.tsx
@@ -15,8 +15,6 @@
 /*
  * Imports.
  */
-/** @jsx jsx */
-import { jsx } from "@emotion/react";
 import { useContext } from "react";
 import StyleContext, { StyleProvider } from "../../../../../styles/styleContext";
 import { defaultPalette } from "../../../../../styles/styles";